--- python/shotgun_model/shotgun_hierarchy_item.py
@@ -1,17 +1,0 @@
-# Copyright (c) 2016 Shotgun Software Inc.
-#
-# CONFIDENTIAL AND PROPRIETARY
-#
-# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
-# Source Code License included in this distribution package. See LICENSE.
-# By accessing, using, copying or modifying this work you indicate your
-# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
-# not expressly granted therein are reserved by Shotgun Software Inc.
-
-from .shotgun_standard_item_base import ShotgunStandardItemBase
-
-
-class ShotgunHierarchyItem(ShotgunStandardItemBase):
-    # XXX docs
-
-    pass
